% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpjr_data.R
\name{dpjr_data}
\alias{dpjr_data}
\title{Get file path to data files}
\source{
This function is adapted from `readxl::readxl_example()` and `palmerpenguins::path_to_file`.
}
\usage{
dpjr_data(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
for example `"mtcars.csv"` will work.
If `NULL`, the example files will be listed.}
}
\description{
dpjr comes bundled with various data files in its `inst/extdata`
directory. This function make them easy to access.
}
\examples{
# return a list of all the files in the package
dpjr_data()
#
# return the full path of the named file
dpjr_data("mtcars.csv")
#
# return the head (first 6 rows) of the named file
head(read.csv(dpjr_data("mtcars.csv")))
}
